# implement new, faster solving algorithm