--- a/benches/lib.rs
+++ b/benches/lib.rs
@@ -19,7 +19,7 @@ fn read_sudokus(sudokus_str: &str) -> Vec<Sudoku> {
 
 #[bench]
 fn easy_sudokus_solve(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("easy_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/easy_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for mut sudoku in sudokus_1000.iter().cloned() { sudoku.solve(); }
@@ -28,7 +28,7 @@ fn easy_sudokus_solve(b: &mut test::Bencher) {
 
 #[bench]
 fn easy_sudokus_solve_unique(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("easy_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/easy_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_unique(); }
@@ -37,7 +37,7 @@ fn easy_sudokus_solve_unique(b: &mut test::Bencher) {
 
 #[bench]
 fn easy_sudokus_solve_one(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("easy_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/easy_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_one(); }
@@ -46,7 +46,7 @@ fn easy_sudokus_solve_one(b: &mut test::Bencher) {
 
 #[bench]
 fn easy_sudokus_solve_at_most_100(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("easy_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/easy_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_at_most(100); }
@@ -55,7 +55,7 @@ fn easy_sudokus_solve_at_most_100(b: &mut test::Bencher) {
 
 #[bench]
 fn medium_sudokus_solve(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("medium_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/medium_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for mut sudoku in sudokus_1000.iter().cloned() { sudoku.solve(); }
@@ -64,7 +64,7 @@ fn medium_sudokus_solve(b: &mut test::Bencher) {
 
 #[bench]
 fn medium_sudokus_solve_unique(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("medium_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/medium_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_unique(); }
@@ -73,7 +73,7 @@ fn medium_sudokus_solve_unique(b: &mut test::Bencher) {
 
 #[bench]
 fn medium_sudokus_solve_one(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("medium_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/medium_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_one(); }
@@ -82,7 +82,7 @@ fn medium_sudokus_solve_one(b: &mut test::Bencher) {
 
 #[bench]
 fn medium_sudokus_solve_at_most_100(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("medium_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/medium_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_at_most(100); }
@@ -91,7 +91,7 @@ fn medium_sudokus_solve_at_most_100(b: &mut test::Bencher) {
 
 #[bench]
 fn hard_sudokus_solve(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("hard_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/hard_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for mut sudoku in sudokus_1000.iter().cloned() { sudoku.solve(); }
@@ -100,7 +100,7 @@ fn hard_sudokus_solve(b: &mut test::Bencher) {
 
 #[bench]
 fn hard_sudokus_solve_unique(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("hard_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/hard_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_unique(); }
@@ -109,7 +109,7 @@ fn hard_sudokus_solve_unique(b: &mut test::Bencher) {
 
 #[bench]
 fn hard_sudokus_solve_one(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("hard_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/hard_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_one(); }
@@ -118,7 +118,7 @@ fn hard_sudokus_solve_one(b: &mut test::Bencher) {
 
 #[bench]
 fn hard_sudokus_solve_at_most_100(b: &mut test::Bencher) {
-    let sudokus = read_sudokus( &include_str!("hard_sudokus.txt") );
+    let sudokus = read_sudokus( &include_str!("../sudokus/hard_sudokus.txt") );
     let sudokus_1000 = sudokus.iter().cycle().cloned().take(100).collect::<Vec<_>>();;
 	b.iter(|| {
 		for sudoku in sudokus_1000.iter().cloned() { sudoku.solve_at_most(100); }