--- a/src/consts.rs
+++ b/src/consts.rs
@@ -1,0 +1,4 @@
+pub const N_SUDOKU_CELLS: usize = 81;
+pub const COL_OFFSET: usize = 81;
+pub const FIELD_OFFSET: usize = 162;
+pub const CELL_OFFSET: usize = 243;