@Emerentius 2016-02-06
I can squash the commits before merge, but I'll keep them distinct for now